consolidate sph kernel tests with fixtures
